requirements.txt
streamlit==1.43.0
langchain==0.3.20
langchain-community==0.3.19
unstructured==0.16.23
sentence-transformers==3.4.1
python-dotenv==1.0.1
groq==0.18.0
faiss-cpu==1.10.0